Modify ingredient endpoints to reflect MinQuantity becoming float

+++ BWBE/Bodies/IngredientInit.cs
@@ -8,6 +8,6 @@ public class IngredientInit
     [MaxLength(36)] public string InventoryId { get; set; } = null!;
     [MaxLength(50)] public string Name { get; set; } = null!;
     public float Quantity { get; set; }
-    public int MinQuantity { get; set; }
+    public float MinQuantity { get; set; }
     [MaxLength(50)] public string Unit { get; set; } = null!;
 }
